refactor : 사다리의 한 라인을 만드는 역할을 분리
역할을 분리해 RandomGenerator 위의 역할 위임
