+++ .idea/workspace.xml
@@ -4,9 +4,13 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="75cfef9c-bc50-442b-99d5-2d9016fefebb" name="Changes" comment="docs : README 작성">
+    <list default="true" id="75cfef9c-bc50-442b-99d5-2d9016fefebb" name="Changes" comment="docs : step4 구현계획&#10;&#10;README 작성">
+      <change afterPath="$PROJECT_DIR$/src/main/java/ladder/domain/RandomGenerator.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/test/java/ladder/domain/RandomGeneratorTest.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/README.md" beforeDir="false" afterPath="$PROJECT_DIR$/README.md" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/ladder/domain/Ladder.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/ladder/domain/Ladder.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/ladder/domain/LadderLine.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/ladder/domain/LadderLine.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -77,10 +81,11 @@
       <recent name="" />
     </key>
     <key name="CreateTestDialog.RecentsKey">
+      <recent name="ladder.domain" />
       <recent name="kr.codesquad" />
     </key>
   </component>
-  <component name="RunManager">
+  <component name="RunManager" selected="JUnit.RandomGeneratorTest.randomLine">
     <configuration name="Main" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
       <option name="MAIN_CLASS_NAME" value="ladder.Main" />
       <module name="java-lotto.main" />
@@ -101,9 +106,60 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
+    <configuration name="PlayersTest.failTest" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
+      <module name="java-lotto.test" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="ladder.domain.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <option name="PACKAGE_NAME" value="ladder.domain" />
+      <option name="MAIN_CLASS_NAME" value="ladder.domain.PlayersTest" />
+      <option name="METHOD_NAME" value="failTest" />
+      <option name="TEST_OBJECT" value="method" />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+    <configuration name="PlayersTest.playerNameCheckTest" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
+      <module name="java-lotto.test" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="ladder.domain.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <option name="PACKAGE_NAME" value="ladder.domain" />
+      <option name="MAIN_CLASS_NAME" value="ladder.domain.PlayersTest" />
+      <option name="METHOD_NAME" value="playerNameCheckTest" />
+      <option name="TEST_OBJECT" value="method" />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+    <configuration name="RandomGeneratorTest.randomLine" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
+      <module name="java-lotto.test" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="ladder.domain.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <option name="PACKAGE_NAME" value="ladder.domain" />
+      <option name="MAIN_CLASS_NAME" value="ladder.domain.RandomGeneratorTest" />
+      <option name="METHOD_NAME" value="randomLine" />
+      <option name="TEST_OBJECT" value="method" />
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="JUnit.RandomGeneratorTest.randomLine" />
         <item itemvalue="Application.Main" />
+        <item itemvalue="JUnit.PlayersTest.failTest" />
+        <item itemvalue="JUnit.PlayersTest.playerNameCheckTest" />
       </list>
     </recent_temporary>
   </component>
@@ -128,6 +184,9 @@
       <workItem from="1678850713355" duration="4590000" />
       <workItem from="1678861987430" duration="1762000" />
       <workItem from="1678864699444" duration="2231000" />
+      <workItem from="1678867298473" duration="685000" />
+      <workItem from="1678889685122" duration="3912000" />
+      <workItem from="1678929470483" duration="13463000" />
     </task>
     <task id="LOCAL-00001" summary="docs : 학습계획 정리">
       <created>1678244485492</created>
@@ -150,7 +209,14 @@
       <option name="project" value="LOCAL" />
       <updated>1678867086193</updated>
     </task>
-    <option name="localTasksCounter" value="4" />
+    <task id="LOCAL-00004" summary="docs : step4 구현계획&#10;&#10;README 작성">
+      <created>1678867382187</created>
+      <option name="number" value="00004" />
+      <option name="presentableId" value="LOCAL-00004" />
+      <option name="project" value="LOCAL" />
+      <updated>1678867382187</updated>
+    </task>
+    <option name="localTasksCounter" value="5" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -171,6 +237,23 @@
     <MESSAGE value="docs : 학습계획 정리" />
     <MESSAGE value="refactor : 접근제어자 수정, 불필요한 주석제거 등&#10;&#10;코딩컨벤션을 준수하기 위한 메서드간의 개행, 빼먹었던 접근제어자 수정, 불필요한 주석 제거" />
     <MESSAGE value="chore : 패키지 분리&#10;&#10;핵심로직과,view 를 분리" />
-    <option name="LAST_COMMIT_MESSAGE" value="chore : 패키지 분리&#10;&#10;핵심로직과,view 를 분리" />
+    <MESSAGE value="docs : step4 구현계획&#10;&#10;README 작성" />
+    <option name="LAST_COMMIT_MESSAGE" value="docs : step4 구현계획&#10;&#10;README 작성" />
+  </component>
+  <component name="XDebuggerManager">
+    <breakpoint-manager>
+      <breakpoints>
+        <line-breakpoint enabled="true" type="java-line">
+          <url>file://$PROJECT_DIR$/src/main/java/ladder/domain/RandomGenerator.java</url>
+          <line>9</line>
+          <option name="timeStamp" value="5" />
+        </line-breakpoint>
+        <line-breakpoint enabled="true" type="java-line">
+          <url>file://$PROJECT_DIR$/src/main/java/ladder/domain/RandomGenerator.java</url>
+          <line>29</line>
+          <option name="timeStamp" value="6" />
+        </line-breakpoint>
+      </breakpoints>
+    </breakpoint-manager>
   </component>
 </project>
